fix: include current shared link expiry date as default select option
