+++ i18n/en.json
@@ -1799,6 +1799,7 @@
   "select_keep_all": "Select keep all",
   "select_library_owner": "Select library owner",
   "select_new_face": "Select new face",
+  "select_option_unchanged": "{label} (unchanged)",
   "select_person_to_tag": "Select a person to tag",
   "select_photos": "Select photos",
   "select_trash_all": "Select trash all",
